move event listener to correct module
