+++ backend/event-listeners/src/domain/mod.rs
@@ -1,5 +1,2 @@
-mod event_listener;
-pub use event_listener::*;
-
 mod github;
 pub use github::Event as GithubEvent;
